Major refactor/re-organization of file structure.
