+++ src/app/pages/movies/movies.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { MoviesService } from "../services/movies.service";
-import { Movie } from "../../models/movie";
+import { MoviesService } from "../../services/movies.service";
+import { Movie } from "../../../models/movie";
 import { ActivatedRoute } from "@angular/router";
 import { take } from "rxjs";
